Changed the name of the variables: lbl_torsRope ->  lblPersonalRope; straightforwardnessRope -> personalRope(textfield); torsRope -> numberRopeMachine(textFields); Some code blocks are commented out due to the removal of some fields from user input Modified style css fxml.

+++ src/main/java/gui/model/TestLabel.java
@@ -29,7 +29,7 @@ public class TestLabel {
     private Integer length;
     private String numberSpool;
     private Integer welds;
-    private Double numberRopeMachine;
+    private Integer numberRopeMachine;
     private Double numberWeldingMachine;
     private Double straightforwardness600_0;
     private Double straightforwardness600_1;
